services/rag-api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
psycopg[binary]==3.2.1
pgvector==0.3.6
pydantic-settings==2.4.0
httpx==0.27.2
